fix(agent): fence RFC-64 immutable persistence races
