+++ .github/workflows/rfc64-inventory-windows.yml
@@ -7,9 +7,9 @@ on:
       - 'packages/agent/src/rfc64/inventory-v1/**'
       - 'packages/agent/src/rfc64/author-catalog-producer.ts'
       - 'packages/agent/src/rfc64/control-object-store-v1.ts'
+      - 'packages/agent/src/rfc64/control-object-store-v1-internal.ts'
       - 'packages/agent/src/rfc64/durable-file-store-v1.ts'
       - 'packages/agent/src/rfc64/persistence-v1.ts'
-      - 'packages/agent/src/rfc64/persistence-owner-capability-v1.ts'
       - 'packages/agent/src/rfc64/secure-filesystem-policy-v1.ts'
       - 'packages/agent/src/dkg-agent-base.ts'
       - 'packages/agent/src/index.ts'
@@ -17,6 +17,7 @@ on:
       - 'packages/agent/test/rfc64-agent-inventory-lifecycle.test.ts'
       - 'packages/agent/test/rfc64-author-catalog-producer.test.ts'
       - 'packages/agent/test/rfc64-control-object-store-v1.test.ts'
+      - 'packages/agent/test/support/rfc64-control-object-store-test-support.ts'
       - 'packages/agent/test/rfc64-control-store-public-api.typecheck.ts'
       - 'packages/agent/test/fixtures/rfc64-inventory-v1-child.ts'
       - 'packages/agent/test/fixtures/rfc64-agent-inventory-lifecycle-child.ts'
