+++ packages/agent/test/rfc64-control-store-public-api.typecheck.ts
@@ -1,49 +1,33 @@
-import * as publicApi from '../src/index.js';
-import {
-  RFC64_CONTROL_OBJECT_STORE_ERROR_CODES_V1,
-  RFC64_CONTROL_OBJECT_STORE_MAX_STAGE_OBJECTS,
-  RFC64_CONTROL_OBJECT_STORE_POSIX_NAMESPACE_DURABILITY,
-  RFC64_CONTROL_OBJECT_STORE_RELATIVE_PATH,
-  RFC64_CONTROL_OBJECT_STORE_WINDOWS_NAMESPACE_DURABILITY,
-  Rfc64ControlObjectStoreErrorV1,
-  type GetVerifiedControlObjectInputV1,
-  type Rfc64ControlObjectStoreErrorCodeV1,
-  type Rfc64ControlObjectStoreNamespaceDurabilityV1,
-  type Rfc64ControlObjectStoreV1,
-  type StageVerifiedControlObjectV1,
-  type StageVerifiedControlObjectsResultV1,
-  type StagedVerifiedControlObjectV1,
-  type StoredVerifiedControlObjectV1,
-} from '../src/index.js';
+import * as packageRoot from '../src/index.js';
+import * as productionControlStoreModule from '../src/rfc64/control-object-store-v1.js';
 
-type PublicApiHasRawControlStoreOpener =
-  'openRfc64ControlObjectStoreV1' extends keyof typeof publicApi ? true : false;
+type PackageRootHasRawControlStoreOpener =
+  'openRfc64ControlObjectStoreV1' extends keyof typeof packageRoot ? true : false;
+type PackageRootHasControlStoreLayout =
+  'RFC64_CONTROL_OBJECT_STORE_RELATIVE_PATH' extends keyof typeof packageRoot
+    ? true
+    : false;
+type ProductionModuleHasRawControlStoreOpener =
+  'openRfc64ControlObjectStoreV1' extends keyof typeof productionControlStoreModule
+    ? true
+    : false;
+type ProductionModuleHasTestOpener =
+  'createRfc64ControlObjectStoreTestOpenerV1' extends
+  keyof typeof productionControlStoreModule ? true : false;
 
-// Package consumers cannot construct the cache without the aggregate
-// persistence owner; direct-source tests retain a guarded internal seam.
-const rawControlStoreOpenerIsNotPublic: PublicApiHasRawControlStoreOpener = false;
+// The stable package root intentionally withholds the implementation-shaped
+// cache API until an RFC-64 public workflow consumes it. Internal dist
+// subpaths are unsupported; this test defines the stable root/module contract.
+const packageRootHasRawControlStoreOpener: PackageRootHasRawControlStoreOpener = false;
+const packageRootHasControlStoreLayout: PackageRootHasControlStoreLayout = false;
+const productionModuleHasRawControlStoreOpener: ProductionModuleHasRawControlStoreOpener = false;
+const productionModuleHasTestOpener: ProductionModuleHasTestOpener = false;
 
-const publicControlStoreRuntimeExports = [
-  RFC64_CONTROL_OBJECT_STORE_ERROR_CODES_V1,
-  RFC64_CONTROL_OBJECT_STORE_MAX_STAGE_OBJECTS,
-  RFC64_CONTROL_OBJECT_STORE_POSIX_NAMESPACE_DURABILITY,
-  RFC64_CONTROL_OBJECT_STORE_RELATIVE_PATH,
-  RFC64_CONTROL_OBJECT_STORE_WINDOWS_NAMESPACE_DURABILITY,
-  Rfc64ControlObjectStoreErrorV1,
-] as const;
+// @ts-expect-error Low-level store types are not part of the stable package root.
+type PackageRootStoreType = packageRoot.Rfc64ControlObjectStoreV1;
 
-type PublicControlStoreTypesAreUsable = readonly [
-  GetVerifiedControlObjectInputV1,
-  Rfc64ControlObjectStoreErrorCodeV1,
-  Rfc64ControlObjectStoreNamespaceDurabilityV1,
-  Rfc64ControlObjectStoreV1,
-  StageVerifiedControlObjectV1,
-  StageVerifiedControlObjectsResultV1,
-  StagedVerifiedControlObjectV1,
-  StoredVerifiedControlObjectV1,
-];
-const publicControlStoreTypesAreUsable: PublicControlStoreTypesAreUsable | undefined = undefined;
-
-void rawControlStoreOpenerIsNotPublic;
-void publicControlStoreRuntimeExports;
-void publicControlStoreTypesAreUsable;
+void packageRootHasRawControlStoreOpener;
+void packageRootHasControlStoreLayout;
+void productionModuleHasRawControlStoreOpener;
+void productionModuleHasTestOpener;
+void (undefined as PackageRootStoreType | undefined);
